Fix for in_udp_accept not being applied.
in_udp_accept did not have the "elements" line needed to actually apply any ports added to the defined list.
Simply adding the line (and probably why it was removed) led to nftables errors though, as "none" is not a valid choice for a define.
To fix this, logic to filter out any empty or 'none' defines was added.

This allows defines to be either empty or in the 'none' state.
For example:
  input udp accepted:
    name: in_udp_accept
    value: 'none'

Signed-off-by: Philipp Rintz <git@rintz.net>