extend nft_flush_target to nft_flush_cmd
When using just a 'target' and you want to preserve
some custom tables you might want to change the behavior of
this role to just handle the 'inet filter' table.

But running 'flush table inet filter' will not remove
the filter table as it is done with the default 'flush ruleset'.
Instead this will just flush the rules and sets and chains would
be preserved and maybe leading to undesired side effects.

This commit extends the variable to hold a complete command
so you can use 'destroy' on a table instead of 'flush' so the
filter table is rebuild from the ground up on restart.